move SignatureHashInterface to Transaction namespace
